Schedule Development
Began working on the 'Schedule' class of the application. This class will include a specific set of dates that will provide times for which an inputted amount of employees will work. The 'Schedule' class will be used in the Manager class when they create one each week. Also created a seperate struct named Time to determine when employees clock in and out. An organizational folder named Utility was also created for the Schedule class and will hold other functional classes for the application. README was also updated on the new classes and their various attributes.