Clock In, Clock Out Update!
Simulation of an employee clocking and clocking out is complete in the back end of things. There is still no implementation for user interaction, obviously there is a long way to go with other aspects of this application before I can begin that process. The Register class has had many additions and the big three have been implemented for both the Register class and Employee class.